+++ src/search/adapters/index.js
@@ -1,7 +1,5 @@
 import { createHtmlAdapter } from "./html.js";
 import { createRssAdapter } from "./rss.js";
-import { createPastetodayAdapter } from "./pastetoday.js";
-import { createOfversedropsAdapter } from "./ofversedrops.js";
 import { createCustomAdapter } from "./custom.js";
 
 const htmlAdapter = createHtmlAdapter();
@@ -9,9 +7,7 @@ const ADAPTERS = new Map([
   ["html", htmlAdapter],
   ["json", htmlAdapter],
   ["custom", createCustomAdapter()],
-  ["rss", createRssAdapter()],
-  ["pastetoday", createPastetodayAdapter()],
-  ["ofversedrops", createOfversedropsAdapter()]
+  ["rss", createRssAdapter()]
 ]);
 
 export function adapterForSource(source) {